SuperChat/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-telegram-bot
requests